Add status on details
