+++ core-ui/src/components/Predefined/Details/Application/Application.details.js
@@ -4,6 +4,7 @@ import { Labels } from 'react-shared';
 import ApplicationServices from './ApplicationServices';
 import NamespaceBindings from './NamespaceBindings';
 import ConnectApplicationModal from './ConnectApplicationModal';
+import { ApplicationStatus } from './ApplicationStatus';
 
 export const ApplicationsDetails = DefaultRenderer => ({ ...otherParams }) => {
   const customColumns = [
@@ -15,7 +16,12 @@ export const ApplicationsDetails = DefaultRenderer => ({ ...otherParams }) => {
       header: 'Labels',
       value: app => <Labels labels={app.metadata.labels} />,
     },
+    {
+      header: 'Status',
+      value: app => <ApplicationStatus application={app} />,
+    },
   ];
+
   return (
     <DefaultRenderer
       customColumns={customColumns}
