+++ core-ui/src/components/Predefined/List/Application/Applications.list.js
@@ -1,31 +1,12 @@
 import React from 'react';
 import CreateApplicationModal from './CreateApplicationModal';
-import { StatusBadge } from 'react-shared';
-
-const isStatusOk = application => {
-  return application.status?.installationStatus?.status === 'deployed';
-};
-const isStatusEmpty = application => !application.status;
-
-const getStatus = application => {
-  if (isStatusEmpty(application)) return 'provisioning';
-  return isStatusOk(application) ? 'deployed' : 'error';
-};
-
-const getStatusType = application => {
-  if (isStatusEmpty(application)) return 'warning';
-  return isStatusOk(application) ? 'success' : 'error';
-};
+import { ApplicationStatus } from '../../Details/Application/ApplicationStatus';
 
 export const ApplicationsList = DefaultRenderer => ({ ...otherParams }) => {
   const customColumns = [
     {
       header: 'Status',
-      value: application => {
-        const status = getStatus(application);
-        const statusType = getStatusType(application);
-        return <StatusBadge type={statusType}>{status}</StatusBadge>;
-      },
+      value: application => <ApplicationStatus application={application} />,
     },
   ];
